= filter autotests by env

+++ plugins/test/autotest.go
@@ -2,6 +2,7 @@ package test
 
 import (
 	"fmt"
+	"log"
 
 	"github.com/InnovaCo/serve/manifest"
 	"github.com/InnovaCo/serve/utils"
@@ -14,6 +15,11 @@ func init() {
 type TestAutotest struct{}
 
 func (p TestAutotest) Run(data manifest.Manifest) error {
+	if data.GetString("env") != data.GetString("current-env") {
+		log.Printf("No test found for `%s`.\n", data.GetString("current-env"))
+		return nil
+	}
+
 	if err := utils.RunCmd("rm -rf tests && git clone --depth 1 --single-branch --recursive %s tests", data.GetString("repo")); err != nil {
 		return fmt.Errorf("Error on clone test git repo: %v", err)
 	}
